PS-3860 Split client to regular (storage token) and manage (manage token)
