+++ tests/Unit/ClientTest.php
@@ -7,301 +7,44 @@ namespace Tests\Keboola\BillingApi\Unit;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Middleware;
-use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 use Keboola\BillingApi\Client;
-use Keboola\BillingApi\Exception\BillingException;
+use Keboola\BillingApi\InternalClient;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\RequestInterface;
-use Psr\Log\Test\TestLogger;
 
 class ClientTest extends TestCase
 {
-    private function getClient(array $options): Client
-    {
-        return new Client(
-            'https://example.com/',
-            'testToken',
-            $options
-        );
-    }
-
-    public function testCreateClientInvalidBackoff(): void
-    {
-        $this->expectException(BillingException::class);
-        $this->expectExceptionMessage(
-            'Invalid parameters when creating client: Value "abc" is invalid: This value should be a valid number'
-        );
-        new Client(
-            'https://example.com/',
-            'testToken',
-            ['backoffMaxTries' => 'abc']
-        );
-    }
-
-    public function testCreateClientTooLowBackoff(): void
-    {
-        $this->expectException(BillingException::class);
-        $this->expectExceptionMessage(
-            'Invalid parameters when creating client: Value "-1" is invalid: This value should be between 0 and 100.'
-        );
-        new Client(
-            'https://example.com/',
-            'testToken',
-            ['backoffMaxTries' => -1]
-        );
-    }
-
-    public function testCreateClientTooHighBackoff(): void
-    {
-        $this->expectException(BillingException::class);
-        $this->expectExceptionMessage(
-            'Invalid parameters when creating client: Value "101" is invalid: This value should be between 0 and 100.'
-        );
-        new Client(
-            'https://example.com/',
-            'testToken',
-            ['backoffMaxTries' => 101]
-        );
-    }
-
-    public function testCreateClientInvalidToken(): void
-    {
-        $this->expectException(BillingException::class);
-        $this->expectExceptionMessage(
-            'Invalid parameters when creating client: Value "" is invalid: This value should not be blank.'
-        );
-        new Client('https://example.com/', '');
-    }
-
-    public function testCreateClientInvalidUrl(): void
-    {
-        $this->expectException(BillingException::class);
-        $this->expectExceptionMessage(
-            'Invalid parameters when creating client: Value "invalid url" is invalid: This value is not a valid URL.'
-        );
-        new Client('invalid url', 'testToken');
-    }
-
-    public function testCreateClientMultipleErrors(): void
-    {
-        $this->expectException(BillingException::class);
-        $this->expectExceptionMessage(
-            'Invalid parameters when creating client: Value "invalid url" is invalid: This value is not a valid URL.'
-        );
-        new Client('invalid url', '');
-    }
-
-    public function testClientRequestResponse(): void
-    {
-        $mock = new MockHandler([
-            new Response(
-                200,
-                ['Content-Type' => 'application/json'],
-                '{
-                    "remaining": "123.4343434343434343",
-                    "consumed": "456.1212121212121212"
-                }'
-            ),
-        ]);
-        // Add the history middleware to the handler stack.
-        $requestHistory = [];
-        $history = Middleware::history($requestHistory);
-        $stack = HandlerStack::create($mock);
-        $stack->push($history);
-        $client = $this->getClient(['handler' => $stack]);
-        $credits = $client->getRemainingCredits();
-        self::assertEqualsWithDelta(123.43434343434, $credits, 0.00001);
-        self::assertCount(1, $requestHistory);
-
-        $request = $requestHistory[0]['request'];
-        self::assertInstanceOf(Request::class, $request);
-        self::assertEquals('https://example.com/credits', $request->getUri()->__toString());
-        self::assertEquals('GET', $request->getMethod());
-        self::assertEquals('testToken', $request->getHeader('X-StorageApi-Token')[0]);
-        self::assertEquals('Billing PHP Client', $request->getHeader('User-Agent')[0]);
-        self::assertEquals('application/json', $request->getHeader('Content-type')[0]);
-    }
-
-    public function testInvalidResponse(): void
-    {
-        $mock = new MockHandler([
-            new Response(
-                200,
-                ['Content-Type' => 'application/json'],
-                'invalid json'
-            ),
-        ]);
-        // Add the history middleware to the handler stack.
-        $requestHistory = [];
-        $history = Middleware::history($requestHistory);
-        $stack = HandlerStack::create($mock);
-        $stack->push($history);
-        $client = $this->getClient(['handler' => $stack]);
-        self::expectException(BillingException::class);
-        self::expectExceptionMessage('Unable to parse response body into JSON: Syntax error');
-        $client->getRemainingCredits();
-    }
-
-    public function testLogger(): void
-    {
-        $mock = new MockHandler([
-            new Response(
-                200,
-                ['Content-Type' => 'application/json'],
-                '{
-                    "remaining": "123",
-                    "consumed": "456"
-                }'
-            ),
-        ]);
-        // Add the history middleware to the handler stack.
-        $requestHistory = [];
-        $history = Middleware::history($requestHistory);
-        $stack = HandlerStack::create($mock);
-        $stack->push($history);
-        $logger = new TestLogger();
-        $client = $this->getClient(['handler' => $stack, 'logger' => $logger, 'userAgent' => 'test agent']);
-        $client->getRemainingCredits();
-
-        $request = $requestHistory[0]['request'];
-        self::assertInstanceOf(Request::class, $request);
-        self::assertEquals('test agent', $request->getHeader('User-Agent')[0]);
-        self::assertTrue($logger->hasInfoThatContains('"GET  /1.1" 200 '));
-        self::assertTrue($logger->hasInfoThatContains('test agent'));
-    }
-
-    public function testRetrySuccess(): void
-    {
-        $mock = new MockHandler([
-            new Response(
-                500,
-                ['Content-Type' => 'application/json'],
-                '{"message" => "Out of order"}'
-            ),
-            new Response(
-                500,
-                ['Content-Type' => 'application/json'],
-                'Out of order'
-            ),
-            new Response(
-                200,
-                ['Content-Type' => 'application/json'],
-                '{
-                    "remaining": "123",
-                    "consumed": "456"
-                }'
-            ),
-        ]);
-        // Add the history middleware to the handler stack.
-        $requestHistory = [];
-        $history = Middleware::history($requestHistory);
-        $stack = HandlerStack::create($mock);
-        $stack->push($history);
-        $client = $this->getClient(['handler' => $stack]);
-        $credits = $client->getRemainingCredits();
-
-        self::assertEquals('123', $credits);
-        self::assertCount(3, $requestHistory);
-        $request = $requestHistory[0]['request'];
-        self::assertInstanceOf(Request::class, $request);
-        self::assertEquals('https://example.com/credits', $request->getUri()->__toString());
-        $request = $requestHistory[1]['request'];
-        self::assertEquals('https://example.com/credits', $request->getUri()->__toString());
-        $request = $requestHistory[2]['request'];
-        self::assertEquals('https://example.com/credits', $request->getUri()->__toString());
-    }
-
-    public function testRetryFailure(): void
-    {
-        $responses = [];
-        for ($i = 0; $i < 30; $i++) {
-            $responses[] = new Response(
-                500,
-                ['Content-Type' => 'application/json'],
-                '{"message" => "Out of order"}'
-            );
-        }
-        $mock = new MockHandler($responses);
-        // Add the history middleware to the handler stack.
-        $requestHistory = [];
-        $history = Middleware::history($requestHistory);
-        $stack = HandlerStack::create($mock);
-        $stack->push($history);
-        $client = $this->getClient(['handler' => $stack, 'backoffMaxTries' => 1]);
-        try {
-            $client->getRemainingCredits();
-            self::fail('Must throw exception');
-        } catch (BillingException $e) {
-            self::assertStringContainsString('500 Internal Server Error', $e->getMessage());
-        }
-        self::assertCount(2, $requestHistory);
-    }
-
-    public function testRetryFailureReducedBackoff(): void
-    {
-        $responses = [];
-        for ($i = 0; $i < 30; $i++) {
-            $responses[] = new Response(
-                500,
-                ['Content-Type' => 'application/json'],
-                '{"message" => "Out of order"}'
-            );
-        }
-        $mock = new MockHandler($responses);
-        // Add the history middleware to the handler stack.
-        $requestHistory = [];
-        $history = Middleware::history($requestHistory);
-        $stack = HandlerStack::create($mock);
-        $stack->push($history);
-        $client = $this->getClient(['handler' => $stack, 'backoffMaxTries' => 3]);
-        try {
-            $client->getRemainingCredits();
-            self::fail('Must throw exception');
-        } catch (BillingException $e) {
-            self::assertStringContainsString('500 Internal Server Error', $e->getMessage());
-        }
-        self::assertCount(4, $requestHistory);
-    }
-
-    public function testRecordJobDuration(): void
+    public function testGetRemainingCredits(): void
     {
         $requestsMade = [];
         $responses = [
-            new Response(200, [], (string) json_encode([
-                'projectId' => 'project-id',
-                'jobId' => 'job-id',
-                'durationSeconds' => 72.7,
-            ])),
+            new Response(200, [], '
+                {
+                    "remaining": "123.4343434343434343",
+                    "consumed": "456.1212121212121212"
+                }
+            '),
         ];
 
         $handlerStack = HandlerStack::create(new MockHandler($responses));
         $handlerStack->push(Middleware::history($requestsMade));
-        $client = new Client('http://example.com', 'dummy-token', [
+        $internalClient = new InternalClient('http://example.com', 'auth-header', 'dummy-token', [
             'handler' => $handlerStack,
         ]);
 
-        $result = $client->recordJobDuration('project-id', 'job-id', 72.7);
+        $client = new Client($internalClient);
+
+        $result = $client->getRemainingCredits();
 
         self::assertCount(1, $requestsMade);
         $request = $requestsMade[0]['request'];
 
         self::assertInstanceOf(RequestInterface::class, $request);
-        self::assertSame('PUT', $request->getMethod());
-        self::assertSame('http://example.com/duration/job', (string) $request->getUri());
-        self::assertSame(
-            json_encode([
-                'projectId' => 'project-id',
-                'jobId' => 'job-id',
-                'durationSeconds' => 72.7,
-            ]),
-            (string) $request->getBody()
-        );
+        self::assertSame('GET', $request->getMethod());
+        self::assertSame('http://example.com/credits', (string) $request->getUri());
+        self::assertSame('dummy-token', $request->getHeaderLine('auth-header'));
 
-        self::assertSame([
-            'projectId' => 'project-id',
-            'jobId' => 'job-id',
-            'durationSeconds' => 72.7,
-        ], $result);
+        self::assertEqualsWithDelta(123.43434343434, $result, 0.00001);
     }
 }
